Rework of the routes, internal errors handling and using linux epoch as timestamp, instead of a specific time format
